# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AdWallet Africa API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /campaigns:
    get:
      operationId: getCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: platform
          schema:
            type: string
      responses:
        "200":
          description: Campaigns list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"
  /campaigns/{id}/launch:
    post:
      operationId: launchCampaign
      tags: [campaigns]
      summary: Launch a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign launched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause an active campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # Analytics
...
components:
  schemas:
...
    MessageResult:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        businessName:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, agency, admin]
        businessName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # Wallet
...
    Campaign:
      type: object
      required: [id, name, platform, objective, status, dailyBudget, totalBudget, creditsUsed, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        platform:
          type: string
          enum: [facebook, instagram, tiktok, google, youtube]
        objective:
          type: string
          enum: [awareness, engagement, traffic, leads, sales]
        status:
          type: string
          enum: [draft, active, paused, completed]
        dailyBudget:
          type: number
        totalBudget:
          type: number
        creditsUsed:
          type: number
        impressions:
          type: ["integer", "null"]
        clicks:
          type: ["integer", "null"]
        conversions:
          type: ["integer", "null"]
        reach:
          type: ["integer", "null"]
        headline:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        targetLocation:
          type: ["string", "null"]
        targetAge:
          type: ["string", "null"]
        targetGender:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        launchedAt:
          type: ["string", "null"]
          format: date-time
    CampaignInput:
      type: object
      required: [name, platform, objective, dailyBudget, totalBudget]
      properties:
        name:
          type: string
        platform:
          type: string
          enum: [facebook, instagram, tiktok, google, youtube]
        objective:
          type: string
          enum: [awareness, engagement, traffic, leads, sales]
        dailyBudget:
          type: number
        totalBudget:
          type: number
        headline:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        targetLocation:
          type: ["string", "null"]
        targetAge:
          type: ["string", "null"]
        targetGender:
          type: ["string", "null"]
    CampaignUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        headline:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        dailyBudget:
          type: ["number", "null"]
        totalBudget:
          type: ["number", "null"]
        targetLocation:
          type: ["string", "null"]
        targetAge:
          type: ["string", "null"]
        targetGender:
          type: ["string", "null"]

    # Analytics
...